Implement a custom handler to allow interaction during the evaluation process
