+++ template/template.go
@@ -218,7 +218,7 @@ func (t *Template) initExtension() {
 
 		// We execute the content, but we ignore errors. The goal is only to register the sub templates and aliases properly
 		// We also do not ask to clone the context as we wish to let extension to be able to alter the supplied context
-		if _, err := ext.processContentInternal(content, file, nil, 0, false); err != nil {
+		if _, _, err := ext.processContentInternal(content, file, nil, 0, false, nil); err != nil {
 			log.Error(err)
 		}
 	}
